Project solution to interview question
Question:

Create a pool of “listener” threads (the number of them passed as a
parameter via the Java command line) that wait on a Queue. Create
another thread that puts random values as “Integer” Java Objects onto
the queue at an interval specified via a Java command line parameter.
When a listener thread grabs an Integer off the Queue, have it print out
the value, along with its Thread name. Hint: the listener thread pool
can be signaled with a specific Java concurrency call.